requirements.txt
# Core dependencies
telethon>=1.36.0
moviepy>=1.0.3
Pillow>=10.3.0
imageio-ffmpeg>=0.4.9

# Google APIs
google-api-python-client>=2.130.0
google-auth>=2.31.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-generativeai>=0.3.0
google-api-core>=2.14.0

# Utilities
python-dotenv>=1.0.1
tenacity>=8.5.0
numpy>=1.24.0
psutil>=5.9.0

# Video Generation V2 (HTML+Selenium)
selenium>=4.15.0
opencv-python>=4.8.0

# Optional: For advanced features (if needed)
# playwright>=1.40.0  # Only if using HTML rendering